attempt to fix more typing

+++ art/attacks/evasion/graphite/graphite_whitebox_pytorch.py
@@ -115,7 +115,7 @@ class GRAPHITEWhiteboxPyTorch(EvasionAttack):
         crop_percent_range: Tuple[float, float] = (-0.03125, 0.03125),
         off_x_range: Tuple[float, float] = (-0.03125, 0.03125),
         off_y_range: Tuple[float, float] = (-0.03125, 0.03125),
-        blur_kernels: Union[Tuple[float, float], List[float]] = (0, 3),
+        blur_kernels: Union[Tuple[int, int], List[int]] = (0, 3),
         batch_size: int = 64,
     ) -> None:
         """
